+++ web/twa-vis-platform/code/src/app/help/page.tsx
@@ -1,14 +1,28 @@
-import styles from './help.module.css';
-
 import React from 'react';
+
+import markdownIt from "markdown-it";
+
+import StaticContentPage from 'ui/pages/static-content-page';
+import OptionalPages, { OptionalPage } from 'io/config/optional-pages';
+
+// Utilities to render markdown into HTML
+const markdowner = markdownIt({
+  html: true,
+  typographer: true,
+  breaks: true,
+  linkify: true
+});
 
 /**
  * Renders the help page.
  */
 export default function HelpPage() {
+  const page: OptionalPage = OptionalPages.getPage("help");
+  const markdownResult = markdowner.render(page.content);
+
   return (
-    <div className={styles.container}>
-      <h1>Help Page is under construction.</h1>
-    </div>
+    <StaticContentPage
+      childString={markdownResult}
+    />
   )
 }